Moved all of game into game.js from index
Create image elements in dinosaur constructor instead of html
