+++ index.html
@@ -20,8 +20,6 @@
         <h3>Bubble Bobble</h3>
         <h5>Click to Start!</h2>
       </div>
-      <img id="dino_right" src="lib/bob_right.png">
-      <img id="dino_left" src="lib/bob_left.png">
     </div>
     <div class="row">
       <div class="col-md-6 scores">
