+++ lib/dinosaur.js
@@ -7,8 +7,8 @@ function Dinosaur(canvas, dino_img_left, dino_img_right) {
   this.y = canvas.height - this.height-10;
   this.status = null;
   this.direction = "right";
-  this.dino_img_left = dino_img_left;
-  this.dino_img_right = dino_img_right;
+  this.dino_img_left = createImage("lib/bob_left.png");
+  this.dino_img_right = createImage("lib/bob_right.png");
   this.count = 0;
   this.canvas = canvas;
   this.points = 0;
@@ -164,6 +164,13 @@ function dontHitCeiling(dino) {
     dino.y = 0;
     dino.count = dino.jumpSteps;
   }
+}
+
+function createImage(imageSrc) {
+  var image = document.createElement("img");
+  image.src = imageSrc;
+  image.style.visibility = 'hidden';
+  return image;
 }
 
 module.exports = Dinosaur;
